moved Console from UI to Logic namespace
