+++ src/utils/Utils.cpp
@@ -457,7 +457,8 @@ bool Utils::containsLike(const std::string& searchSpace, const std::string& part
 }
 
 
-UI::ConsoleCommand::Suggestion Utils::findSuggestion(const std::vector<UI::ConsoleCommand::Suggestion>& suggestions, const std::string& name){
+Logic::ConsoleCommand::Suggestion Utils::findSuggestion(const std::vector<Logic::ConsoleCommand::Suggestion>& suggestions,
+                                                        const std::string& name){
     for (auto& suggestion : suggestions)
     {
         auto& aliasList = suggestion->aliasList;
